+++ Final_Project/Projkt - Kaa/Assets/Prefabs/platforms_galaxy.prefab
@@ -34,7 +34,7 @@ Transform:
   m_GameObject: {fileID: 6261910902521260395}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 6.519, y: 2.163, z: 0}
-  m_LocalScale: {x: 0.4743856, y: 0.2860063, z: 1}
+  m_LocalScale: {x: 0.3743856, y: 0.2860063, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
