+++ Final_Project/Projkt - Kaa/Assets/Scenes/GameSceneNight.unity
@@ -2716,9 +2716,9 @@ MonoBehaviour:
   Tail: {fileID: 1599537421173734056, guid: 5bef62e65a2cc8b44abf7cbda9316061, type: 3}
   TailOrigin: {fileID: 1541840501}
   TempPoint: {fileID: 328130652}
-  QuickFireCooldown: 2.5
-  Ability1Cooldown: 1.3
-  Ability2Cooldown: 1.4
+  QuickFireCooldown: 0.2
+  Ability1Cooldown: 3.5
+  Ability2Cooldown: 1
 --- !u!95 &1342263478
 Animator:
   serializedVersion: 3
@@ -3452,7 +3452,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9af157acb9687754a9ac24589b16ddc8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  PortalWait: 10
+  PortalWait: 75
   portal: {fileID: 3583833958037535332, guid: 264fecf6a88857040952cf8d155f647e, type: 3}
   PortalLocation: {fileID: 1763086097}
   PortalPoint: {fileID: 1796088089}
