+++ Final_Project/Projkt - Kaa/Assets/Scripts/.vs/VSWorkspaceState.json
@@ -2,6 +2,6 @@
   "ExpandedNodes": [
     ""
   ],
-  "SelectedNode": "\\PortalMove.cs",
+  "SelectedNode": "\\Player2AbilityManager.cs",
   "PreviewInSolutionExplorer": false
 }
