--- Final_Project/Projkt - Kaa/Assets/Scripts/Player2QuickFire.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac919240c6b66374ba8eb69c79e1a568
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
