+++ Final_Project/Projkt - Kaa/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/StartScene.unity
+- path: Assets/Scenes/GameSceneDay.unity
   isLoaded: 1
   isActive: 1
