+++ Final_Project/Projkt - Kaa/Library/StateCache/Hierarchy/4ce7b5-5bda54f82df988b4fb.json
@@ -1,1 +1,1 @@
-﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[],"m_ScrollY":0.0,"m_LastClickedFileID":6261910902521260395,"m_LastClickedInstanceID":0}
+﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[],"m_ScrollY":0.0,"m_LastClickedFileID":0,"m_LastClickedInstanceID":0}
