+++ Final_Project/Projkt - Kaa/Library/StateCache/SceneView/8bc60f-3272b8338d706fe41a.json
@@ -1,0 +1,1 @@
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":6.519000053405762,"y":2.1630001068115236,"z":0.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":4.6912431716918949,"orthographic":true}
