+++ Final_Project/Projkt - Kaa/Library/StateCache/SceneView/8bc60f-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":6.886453628540039,"y":1.4767860174179078,"z":0.25632601976394656},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":20.471965789794923,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":-11.83969783782959,"y":-1.2620187997817994,"z":0.24609375},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":29.938961029052736,"orthographic":true}
